I was able to fetch the update items and pand pass them to the html dom. Up next is to do the update in the database

+++ app.js
@@ -3,6 +3,10 @@ const date=document.querySelector('#startDate');
 const sub=document.querySelector('.user-input');
 const submitted=document.querySelector('.new-input');
 const gettodos=document.querySelector('.todos');
+const newform=document.querySelector('.users');
+const myitems=document.querySelector('.myitems');
+const mytodo = document.querySelector('.inner-todos');
+
 
 if(sub){
 sub.addEventListener('submit', e =>{
@@ -10,18 +14,20 @@ sub.addEventListener('submit', e =>{
     console.log("click go here");
     console.log(sub.activity.value);
     console.log(sub.exampleFormControlTextarea1.value.trim());
-    newDate=(dateFns.format(sub.startDate.value.trim(), 'MMMM d, YYYY'));
     const now = new Date();
     todoDate=sub.startDate.value;
+    console.log(todoDate);
     
     const todo={
         Activity:sub.activity.value,
         Created_at:firebase.firestore.Timestamp.fromDate(now),
-        DateofActivity:newDate,
+        DateofActivity:todoDate,
         Description:sub.exampleFormControlTextarea1.value,
         Status:"Pending"
     };
-    db.collection('todos').add(todo).then(()=>{
+    console.log(todo);
+    db.collection('todos').add(todo).then((result)=>{
+        console.log(result);
     }).catch(err =>{
         console.log(err);
     });
@@ -29,20 +35,24 @@ sub.addEventListener('submit', e =>{
 });
 }
 
+
 const display=(todo, id)=>{
     let time = todo.Created_at.toDate();
     let html=`<li data-id="${id}">
-    <div><h3>Activity</h3>${todo.Activity}</div>
-    <div><h3>Description</h3>${todo.Description}</div>
-    <div><h3>Status</h3>${todo.Status}</div>
-    <div><h3>Execution Date</h3>${todo.DateofActivity}</div>
+    <div class="update"><h3>Activity</h3>${todo.Activity}</div>
+    <div class="update"><h3>Description</h3>${todo.Description}</div>
+    <div class="update"><h3>Status</h3>${todo.Status}</div>
+    <div class="update"><h3>Execution Date</h3>${todo.DateofActivity}</div>
     <div><h3>Created Date</h3>${time}</div>
     <button class ="btn btn-danger btn-sm my-2">Delete</button>
     <button class ="btn btn-secondary btn-sm my-2">update</button>
-
+    <button class ="btn btn-secondary btn-sm my-2 save">save</button>
     </li>`;
     gettodos.innerHTML+=html;
 }
+
+
+
 const deletetodo=(id)=>{
     const items =document.querySelectorAll('li');
     items.forEach(item=>{
@@ -57,7 +67,9 @@ db.collection('todos').onSnapshot(snapshot=>{
     snapshot.docChanges().forEach(change=>{
         const doc=change.doc;
         if(change.type==='added'){
+            if(gettodos){
             display(doc.data(), doc.id);
+            }
         }
         else if(change.type === 'removed')
         {
@@ -92,8 +104,9 @@ db.collection('todos').onSnapshot(snapshot=>{
 //         }).catch((err)=>{
 //             console.log(err);
 //     });
-// }
+//}
 
+// fetchdata();
 
 const presentNewDate=()=>{
     console.log("this function was called");
@@ -116,17 +129,35 @@ function pageRedirect(){
     },delay);
  }
 
- gettodos.addEventListener('click', e =>{
-    if(e.target.tagName ==='BUTTON')
-    {
-        if(e.target.childNodes[0].data === 'Delete'){
-            console.log("delete got clicked");
-            const id=e.target.parentElement.getAttribute('data-id');
-            db.collection('todos').doc(id).delete();
+ arrayelement=["Activity", "Description","Status","carryout_date"]
+
+
+if (gettodos) {
+    gettodos.addEventListener('click', e =>{
+        if(e.target.tagName ==='BUTTON')
+        {
+            if(e.target.childNodes[0].data === 'Delete'){ 
+                console.log("delete got clicked");
+                const id=e.target.parentElement.getAttribute('data-id');
+                db.collection('todos').doc(id).delete();
+            }
+            else if (e.target.childNodes[0].data === 'update'){
+                const items=document.querySelector('li');
+                const myinitems=(items.getElementsByTagName('div'));
+                for(let i=0,j =0; i<myinitems.length, i<arrayelement.length; i++, j++)
+                {
+                    arrayelement[i]=myinitems[i].firstChild.nextSibling.data
+                }
+                console.log(arrayelement[1]);
+                mytodo.remove();
+                newform.style.setProperty('display','block');
+                sub.activity.value=(arrayelement[1]);
+                myitems.innerText="Update your todos";
+                console.log(items.getAttribute('data-id'));
+
+            }
         }
-        // else if (e.target.childNodes[0].data === 'update'){
-        //     const item=document.querySelector('li');
-        //     console.log(item.childNodes);
-        // }
-    }
-});
+    });
+ }
+
+
